+++ src/main/java/kitchenpos/dto/request/MenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 import com.sun.istack.NotNull;
