+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.validation.Valid;
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.domain.menu.MenuGroup;
-import kitchenpos.dto.MenuGroupRequest;
+import kitchenpos.dto.request.MenuGroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
