+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -5,8 +5,8 @@ import java.util.List;
 import javax.validation.Valid;
 import kitchenpos.application.OrderService;
 import kitchenpos.domain.order.Order;
-import kitchenpos.dto.OrderRequest;
-import kitchenpos.dto.OrderStatusRequest;
+import kitchenpos.dto.request.OrderRequest;
+import kitchenpos.dto.request.OrderStatusRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
