+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.validation.Valid;
 import kitchenpos.application.TableService;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.dto.OrderTableRequest;
+import kitchenpos.dto.request.OrderTableRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
